move CPU funcs
