+++ MANIFEST.in
@@ -61,6 +61,8 @@ include psutil/arch/freebsd/sensors.c
 include psutil/arch/freebsd/sensors.h
 include psutil/arch/freebsd/sys_socks.c
 include psutil/arch/freebsd/sys_socks.h
+include psutil/arch/netbsd/cpu.c
+include psutil/arch/netbsd/cpu.h
 include psutil/arch/netbsd/mem.c
 include psutil/arch/netbsd/mem.h
 include psutil/arch/netbsd/socks.c
