+++ psutil/arch/netbsd/specific.h
@@ -17,13 +17,9 @@ char *psutil_get_cmd_args(pid_t pid, size_t *argsize);
 //
 PyObject *psutil_get_cmdline(pid_t pid);
 PyObject *psutil_proc_threads(PyObject *self, PyObject *args);
-PyObject *psutil_virtual_mem(PyObject *self, PyObject *args);
-PyObject *psutil_swap_mem(PyObject *self, PyObject *args);
 PyObject *psutil_proc_num_fds(PyObject *self, PyObject *args);
 PyObject *psutil_proc_connections(PyObject *self, PyObject *args);
-PyObject *psutil_per_cpu_times(PyObject *self, PyObject *args);
 PyObject* psutil_disk_io_counters(PyObject* self, PyObject* args);
 PyObject* psutil_proc_exe(PyObject* self, PyObject* args);
 PyObject* psutil_proc_num_threads(PyObject* self, PyObject* args);
-PyObject* psutil_cpu_stats(PyObject* self, PyObject* args);
 PyObject *psutil_proc_cwd(PyObject *self, PyObject *args);
